support of multiple users

+++ src/Controller/ChapterDeleteController.php
@@ -3,9 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Chapter;
-use App\Entity\Project;
-use App\Form\ChapterType;
-use App\Repository\ChapterRepository;
+use App\Security\Voter\ProjectVoter;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -18,15 +16,19 @@ class ChapterDeleteController extends AbstractController
 {
     /**
      * @Route("/{chapter}/delete", name="chapter_delete", methods={"DELETE"})
+     * @param Request $request
+     * @param Chapter $chapter
+     * @return Response
      */
-    public function delete(Request $request, Project $project, Chapter $chapter): Response
+    public function delete(Request $request, Chapter $chapter): Response
     {
+        $this->denyAccessUnlessGranted(ProjectVoter::PROJECT_EDIT, $chapter->getProject());
         if ($this->isCsrfTokenValid('delete'.$chapter->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($chapter);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('project_show', ['project' => $project->getId()]);
+        return $this->redirectToRoute('project_show', ['project' => $chapter->getProject()->getId()]);
     }
 }
